+++ inc/wordpress-cleanup.php
@@ -63,6 +63,9 @@ add_filter( 'body_class', 'ea_singular_body_class' );
 		'sidebar-content',
 		'content-width',
 		'landing-page',
+        'wp-custom-logo',
+        'first-block-cover-full',
+        'first-block-align-full',
 	];
 
 	return array_intersect( $classes, $allowed_classes );